fix authenticated user unrestricted file upload on profile picture edit #1401 make variable names more descriptive create functions for common tasks fix possible sql injection move image sanitization function to library/sanitize.inc.php

+++ interface/patient_file/summary/demographics.php
@@ -1693,64 +1693,53 @@ if (isset($_FILES["profile_picture"])) {
   //size check done.
   //extension check done.
   //if any validation needed be added, please add it below.
-  $bool = 0;
+  
+  $image_verified = false;
+  $user_image_absent = true;
+  $extensions = array("jpg", "png", "jpeg"); 
   $target_file =  basename($_FILES["profile_picture"]["name"]);
-  $imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
-  $verify_image = getimagesize($_FILES["profile_picture"]["tmp_name"]);
-  if($verify_image) {
-    $mime = $verify_image["mime"];
-    $mime_types = array('image/png',
-                            'image/jpeg',
-                            'image/gif',
-                            'image/bmp',
-                            'image/vnd.microsoft.icon');
-    $extensions = array("jpg", "png", "jpeg");
-    //mime check with all image formats.
-    if (in_array($mime, $mime_types)) {
-          $bool = 1;
-        if (in_array($imageFileType, $extensions)) {
-        //if mime type matches, then do a size check
-        //size check
-          if ($_FILES["profile_picture"]["size"] > 20971520) {
-            $bool = 0;
-          }
-          else {
-            $bool = 1;
-          }
+  $image_file_type = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
+  $image_size = $_FILES["profile_picture"]["size"];
+  $image_properties = getimagesize($_FILES["profile_picture"]["tmp_name"]);
+  if($image_properties) {
+    //if image mime type matches, then check file extension, then  check size
+    if (image_has_right_mime($image_properties) && image_has_right_extension($image_file_type, $extensions) && 
+      image_has_right_size($image_size) ) {
+
+          $image_verified = true;//image verification passed     
+    }
+    
+  }
+
+ // $query = "SELECT distinct(group_name) FROM layout_options WHERE form_id = ? ORDER BY group_name";
+//$res = sqlStatement($query, array($_GET['layout_id']));
+
+//while ($row = sqlFetchArray($res)) {
+  
+  if($image_verified) {
+    // check if there is a old image for the patient, if yes then delete it
+    $sql = "SELECT picture_url FROM patient_data WHERE pid = ?";
+    $query = sqlStatement($sql, array($pid));
+    $arr = sqlFetchArray($query);
+    if ($arr['picture_url']) {
+      $url = $arr['picture_url'];
+      // a old image exists, so lets delete it
+      if (unlink($GLOBALS['OE_SITES_BASE']."/".$_SESSION['site_id']."/profile_pictures/".$url)) {
+        $user_image_absent = true;
       }
       else {
-        $bool = 0;
+        //if the image does not delete due to file permissions then the new image wont go.
+        $user_image_absent = false;
       }
     }
-    else {
-      $bool = 0;
-    }
-
   }
-  else {
-        $bool = 0;
-  }
-  // check if there is a old image for the patient, if yes then delete it
-  $sql = "SELECT picture_url FROM patient_data WHERE pid = $pid";
-  $query = sqlQ($sql);
-  $arr = sqlFetchArray($query);
-  if ($arr['picture_url']) {
-    $url = $arr['picture_url'];
-    // a old image exists, so lets delete it
-    if (unlink($GLOBALS['OE_SITES_BASE']."/".$_SESSION['site_id']."/profile_pictures/".$url)) {
-      $bool = true;
-    }
-    else {
-      //if the image does not delete due to file permissions then the new image wont go.
-      $bool = false;
-    }
-  }
+  
   $picture_url = "";
   //begin file uploading
   $destination_directory = $GLOBALS['OE_SITES_BASE']."/".$_SESSION['site_id']."/profile_pictures/";
-  if ($bool) {
-    if (move_uploaded_file($_FILES["profile_picture"]["tmp_name"], $destination_directory.$pid.".".$imageFileType)) {
-        $picture_url = $pid.".".$imageFileType;
+  if ($image_verified && $user_image_absent) {
+    if (move_uploaded_file($_FILES["profile_picture"]["tmp_name"], $destination_directory.$pid.".".$image_file_type)) {
+        $picture_url = $pid.".".$image_file_type;
     }
     else {
       //may be failed due to directory permissions.
@@ -1762,34 +1751,31 @@ if (isset($_FILES["profile_picture"])) {
 }
 
 if ($picture_url) {
-    $imageUrl = "../../../sites/".$_SESSION['site_id']."/profile_pictures/".$picture_url;
-    $successMsg = xl("Profile picture has been set successfully");
-    $errorMsg = xl("Error in setting the profile picture");
   //show success message and update the value in db and also refresh the page.
   if (sqlQ("UPDATE patient_data SET picture_url='$picture_url' WHERE pid='$pid'")) {
         echo "<script>
-    iziToast.success({
-        title: 'User Profile Updated',
-        icon: 'fa fa-picture-o',
-        message: '$successMsg',
-        onOpening: function () {
-          $('#prof_img').attr('src', '$imageUrl');
-      }
-    });
+    if (confirm('profile picture has been set successfully')) {
+      window.location = 'demographics.php?set_pid=$pid';
+    }
+    else {
+       window.location = 'demographics.php?set_pid=$pid';
+    }
     </script>";
   }
   else {
-      echo "<script>
-            iziToast.error({
-                title: 'Error',
-                icon: 'fa fa-warning',
-                message: '$errorMsg',
-            });
+            echo "<script>
+    if (confirm('Error in setting the profile picture')) {
+      window.location = 'demographics.php?set_pid=$pid';
+    }
+    else {
+       window.location = 'demographics.php?set_pid=$pid';
+    }
     </script>";
   }
 }
 else {
   //show failure message.
 }
+
 
 ?>
